modify snn type from Integer to long

+++ src/main/java/com/example/childcareHelp/DAO/BabysitterDAO.java
@@ -28,7 +28,7 @@ public class BabysitterDAO {
         return babysitterRepository.findBabysitterByCondition(babysitter);
     }
 
-    public Optional<Babysitter> getBabysitterBySnn(Integer snn) {
+    public Optional<Babysitter> getBabysitterBySnn(long snn) {
         return babysitterRepository.findById(snn);
     }
     public Babysitter getBabysitter(String email, String password) {
