+++ src/main/java/com/example/childcareHelp/service/BabysitterService.java
@@ -29,7 +29,7 @@ BabysitterService {
         return babysitterDAO.getAllBabysitters();
     }
 
-    public Optional<Babysitter> getBabysitter(Integer snn) {
+    public Optional<Babysitter> getBabysitter(long snn) {
         return babysitterDAO.getBabysitterBySnn(snn);
     }
